refactor[Post] : make new domain package for city
and remove directories inside domain folder
