--- src/main/java/com/codeit/blob/post/repository/CityJpaRepository.java
@@ -1,9 +1,0 @@
-package com.codeit.blob.post.repository;
-
-import com.codeit.blob.post.domain.entity.City;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CityJpaRepository extends JpaRepository<City, Long> {
-}
