+++ src/main/java/com/codeit/blob/post/repository/PostImageJpaRepository.java
@@ -1,6 +1,6 @@
 package com.codeit.blob.post.repository;
 
-import com.codeit.blob.post.domain.entity.PostImage;
+import com.codeit.blob.post.domain.PostImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
