+++ src/main/java/com/codeit/blob/post/repository/PostJpaRepository.java
@@ -1,6 +1,6 @@
 package com.codeit.blob.post.repository;
 
-import com.codeit.blob.post.domain.entity.Post;
+import com.codeit.blob.post.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
